moved L1Metrics to metrics module
