+++ state-reconstruct-fetcher/src/lib.rs
@@ -2,6 +2,6 @@
 #![feature(iter_next_chunk)]
 pub mod constants;
 pub mod l1_fetcher;
-pub mod perf_metric;
+pub mod metrics;
 pub mod snapshot;
 pub mod types;
